refactor scripts

+++ infrastructure/scripts/ws-eks-cleanup-app.sh
@@ -2,11 +2,13 @@ set -e
 
 APP_NAME=${1:-"unicorn-store-spring"}
 
-echo Clean up
+echo Cleaning up $APP_NAME ...
 
 docker images --format "{{.Repository}}:{{.Tag}}" | grep $APP_NAME | xargs -r docker rmi
 for x in `aws ecr list-images --repository-name $APP_NAME --query 'imageIds[*][imageDigest]' --output text`; do aws ecr batch-delete-image --repository-name unicorn-store-spring --image-ids imageDigest=$x; done
 for x in `aws ecr list-images --repository-name $APP_NAME --query 'imageIds[*][imageDigest]' --output text`; do aws ecr batch-delete-image --repository-name unicorn-store-spring --image-ids imageDigest=$x; done
-kubectl delete -f ~/environment/$APP_NAME/k8s
 
-rm -rf ~/environment/$APP_NAME/k8s
+if [ -d ~/environment/$APP_NAME/k8s ]; then
+    kubectl delete -f ~/environment/$APP_NAME/k8s
+    rm -rf ~/environment/$APP_NAME/k8s
+fi
